Disable rebuilding of the manual page database on package install for pbuilder
As noted in http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=561539, not
rebuilding mandb on package installation for pbuilder will speed build times
and not affect package builds. This commit adds the hook to implement this
performance improvement. It also adds the HOOKDIR variable to the pbuilderrc
template. It also updates the spec tests for the setup::cows class.

This also updates the file_on_disk define to take an optional mode (it defaults
to 0644 to preserve previous behavior). The hook uses the defined type but
needs to be executable.