Update debbuilder module to support power, arm
Currently the puppetlabs-debbuilder module only supports defining cows for i386
and amd64. To expand possibilities to building for powerand arm, we have to
make things a bit more flexible. This commit updates the cow_exec.pp manifest
to include execs/crons for i386 cows on i386 and amd64, and then for various
iterations of ppc and arm, those arches.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>