Allow upstream mirrors for pbuilderrc to be configurable
This commit adds configurability of the debian and ubuntu mirrors to use when
creating the pbuilderrc file. This way, if you have a custom internal debian
mirror, you can supply it in your debbuilder inclusion and use it to satisfy
deps for cows instead of the official debian/ubuntu mirrors.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>