Enable setting of OTHERMIRROR in pbuilderrc
This commit updates the debbuilder module to allow the specification of a
different extra mirror than "apt.puppetlabs.com" which is set up by default.
This enables one to set up an internal extra apt repository that they want to
use to satisfy build dependencies, e.g. if one wanted to still use upstream
debian but also extra internal packages. Since this means we won't always be
using apt.puppetlabs.com as the OTHERMIRROR, we may not always want to install
the Puppet Labs keyring. As such, we make this configurable as well (because we
may still want the keyring, even if we don't want the repo).

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>