Enable configuration of components to use from repos
This commit enables one to specify the exact components from a repository to
use for dependency resolution. Previously, we still just assumed all of the
components from the default upstream existed, even if you specified a custom
upstream. This commit also corrects the setting of the MIRRORSITE variable in
the pbuilderrc. Previously, even when setting the debian or ubuntu upstream
mirrors, the subpathing of the url was assumed to be that of the default
upstream. Now we assume when you pass these values in, you really mean it.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>