(RE-2129) Only use one OTHERMIRROR in pbuilderrc
Previously, when setting PE_VER to enable a PE cow, we would continue to
include the apt.puppetlabs.com repo, even though we would use none of
those packages. This commit updates the pbuilderrc to set OTHERMIRROR to
be one or the other, but never both. This has the added bonus of
negating an edge case segfault we've seen on precise-amd64 cows that
have both repos enabled.