(RE-1736) Hack around cumulus cl-basefiles bug in debootstrap
cl-basefiles tries to restart openssh-server during the clinit start of
its init script. openssh-server is not listed in its dependencies, so
this fails, causing the cow itself to fail to build. This commit works
around that problem by manually installing openssh-server during the
debootstrap process, immediately after the core packages are installed
and before the 'required' packages like cl-basefiles are installed.