(maint) Make scripts symlinks
All of the scripts provided here are copied from the debootstrap package
available on Debian 8 (Jessie). All Ubuntu scripts are symlinks of the
lucid script, and all Debian scripts are symlinks of the squeeze script.
CumulusLinux was originally a copy of the wheezy script, so here it has
been make a symlink of the squeeze script.