(maint) Add support for PC1 repo
We're shifting how we're delivering software to focus on a new repo
strategy, namely "puppet collections". This commit enables the PC1 repo
on all platforms where it's available (everywhere but lucid), and
disables the old style (main dependencies) on all new platform
additions, as we will not be creating these old repos when we add support
for new platforms.