(RE-4609) Do not enable PC1 repos in the pbuilderrc
Prior to this commit, I got over-ambitious and enabled the PC1 repos in
our cows. This was unnecessary since we already have the ability to
adhoc enable any given repo for a given build. Since the PC* repos will
be transient, it is more streamlined to use this external functionality
to access the PC* repos when needed. This commit pulls out the logic
that was used to add the PC1 repos to given platforms.