(maint) Actually pass in the mirror vars
This commit is a follow up to a previous commit. We need the ability to
customize the mirrors that we are using in our cows. We have the mirror
variables, but they never actually make it to the manifest that parses
the template that defines the repos the cows hit. This commit adds those
variables to the method that sets up the cows