(maint) Symlinks need to point to files that exist
Prior to this commit, all debian/ubuntu cows require a script to build.
This script was a single copy for an older debian/ubuntu build, but it
was recently switched to a more generic name, either 'debian' or
'ubuntu', so when the older platforms reach EOL, removing them will be a
lot simpler. This is great, unless the 'debian' and 'ubuntu' symlink
sources don't actually exist